diemphuoc_16/09/2021

+++ Book Trading/src/main/java/com/example/springboot/repository/StoreDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StoreDAO extends JpaRepository<Store, Long> {
+public interface StoreDAO extends JpaRepository<Store, Integer> {
 }
